Created simple view for displaying items in fake repo.

+++ MachineSpecs/Models/Computer.cs
@@ -8,8 +8,8 @@ namespace MachineSpecs.Models
     public class Computer
     {
         public int ComputerID { get; set; }
-        public int Memory { get; set; }
-        public int StorageCapacity { get; set; }
+        public decimal Memory { get; set; }
+        public decimal StorageCapacity { get; set; }
         public string StorageType { get; set; }
         public decimal Weight { get; set; }
         public int Power { get; set; }
